fix: add services fix
